v1 (initial version)
This code is working, however, to continue development it needs to be modified to be more efficient (ie. reduce the number of global variables, create libraries, etc..). Reason for this is that when I try to add more functions I believe I hit the SRAM limit, as the serial.print is behaving very oddly.